KDR001 ThanhPV fix bug #108184, #108185

+++ nts.uk/uk.at/at.ctx/function/nts.uk.ctx.at.function.infra/src/main/java/nts/uk/ctx/at/function/infra/generator/holidaysremaining/HolidaysRemainingReportGeneratorImp.java
@@ -395,7 +395,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 					continue;
 				}
 				// E2_3 当月より前, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
-				cells.get(firstRow, 10 + totalMonth).setValue(item.getUsedDays()!=null && item.getUsedDays().intValue()!=0? item.getUsedDays(): null);
+				cells.get(firstRow, 10 + totalMonth).setValue(item.getUsedDays()!=null && item.getUsedDays()!=0? item.getUsedDays(): null);
 				// E3_3 当月より前
 				cells.get(firstRow + 1, 10 + totalMonth).setValue(item.getRemainingDays());
 				if (item.getRemainingDays() < 0) {
@@ -417,7 +417,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 					continue;
 				}
 				// E2_3 当月以降, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
-				cells.get(firstRow, 10 + totalMonth).setValue(item.getMonthlyUsageDays()!=null && item.getMonthlyUsageDays().intValue()!=0? item.getMonthlyUsageDays(): null);
+				cells.get(firstRow, 10 + totalMonth).setValue(item.getMonthlyUsageDays()!=null && item.getMonthlyUsageDays()!=0? item.getMonthlyUsageDays(): null);
 				// E3_3 当月以降
 				if(item.getYearMonth().equals(currentMonth)){
 					cells.get(firstRow + 1, 10 + totalMonth).setValue(item.getMonthlyRemainingDays());
@@ -494,7 +494,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 			cells.get(firstRow, 8).setValue(reserveHolidayImported.getUndigestNumber());
 			
 			if (reserveHolidayImported.getUndigestNumber() != null
-					&& reserveHolidayImported.getUndigestNumber().intValue() > 0) {
+					&& reserveHolidayImported.getUndigestNumber() > 0) {
 				setForegroundRed(cells.get(firstRow, 8));
 			}
 
@@ -525,7 +525,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 				// H2_3 当月より前, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 				cells.get(firstRow, 10 + totalMonth)
 						.setValue(reservedYearHolidayItem.getUsedDays() != null
-								&& reservedYearHolidayItem.getUsedDays().intValue() == 0 ? null: reservedYearHolidayItem.getUsedDays());
+								&& reservedYearHolidayItem.getUsedDays() == 0 ? null: reservedYearHolidayItem.getUsedDays());
 				// H2_4 当月より前
 				cells.get(firstRow + 1, 10 + totalMonth).setValue(reservedYearHolidayItem.getRemainingDays());
 				if (reservedYearHolidayItem.getRemainingDays() < 0) {
@@ -547,7 +547,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 				}
 				// H2_3 当月以降, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 				cells.get(firstRow, 10 + totalMonth).setValue(rsvLeaUsedCurrentMonItem.getUsedNumber() != null 
-						&& rsvLeaUsedCurrentMonItem.getUsedNumber().intValue()==0? null: rsvLeaUsedCurrentMonItem.getUsedNumber());
+						&& rsvLeaUsedCurrentMonItem.getUsedNumber()==0? null: rsvLeaUsedCurrentMonItem.getUsedNumber());
 				if (currentMonth.compareTo(rsvLeaUsedCurrentMonItem.getYearMonth()) != 0) {
 					continue;
 				}
@@ -650,14 +650,14 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 					// I2_3 代休_発生_日数, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 					cells.get(firstRow, 10 + totalMonth)
 							.setValue(statusHolidayItem.getOccurrenceDays() != null
-									&& statusHolidayItem.getOccurrenceDays().intValue() == 0 ? null: statusHolidayItem.getOccurrenceDays());
+									&& statusHolidayItem.getOccurrenceDays() == 0 ? null: statusHolidayItem.getOccurrenceDays());
 					// I3_3 代休_使用_日数, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
-					cells.get(firstRow + 1, 10 + totalMonth).setValue(statusHolidayItem.getUseDays()!=null && statusHolidayItem.getUseDays().intValue()==0? null: statusHolidayItem.getUseDays());
+					cells.get(firstRow + 1, 10 + totalMonth).setValue(statusHolidayItem.getUseDays()!=null && statusHolidayItem.getUseDays()==0? null: statusHolidayItem.getUseDays());
 					if (isRepresentSubstitute) {
 						// I4_3 代休_未消化_日数, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 						cells.get(rowIndexRepresentSubstitute, 10 + totalMonth)
-								.setValue(statusHolidayItem.getUnUsedDays()!=null && statusHolidayItem.getUnUsedDays().intValue()==0? null: statusHolidayItem.getUnUsedDays());
-						if(statusHolidayItem.getUnUsedDays() != null && statusHolidayItem.getUnUsedDays().intValue() > 0) {
+								.setValue(statusHolidayItem.getUnUsedDays()!=null && statusHolidayItem.getUnUsedDays()==0? null: statusHolidayItem.getUnUsedDays());
+						if(statusHolidayItem.getUnUsedDays() != null && statusHolidayItem.getUnUsedDays() > 0) {
 							setForegroundRed(cells.get(rowIndexRepresentSubstitute, 10 + totalMonth));					
 						}
 						
@@ -685,18 +685,18 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 						// I2_3 代休_発生_日数, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 						cells.get(firstRow, 10 + totalMonth)
 								.setValue(currentHolidayItem.getMonthOccurrence() != null
-										&& currentHolidayItem.getMonthOccurrence().intValue() == 0 ? null: currentHolidayItem.getMonthOccurrence());
+										&& currentHolidayItem.getMonthOccurrence() == 0 ? null: currentHolidayItem.getMonthOccurrence());
 						// I3_3 代休_使用_日数, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 						cells.get(firstRow + 1, 10 + totalMonth).setValue(currentHolidayItem.getMonthUse()!=null 
-								&& currentHolidayItem.getMonthUse().intValue()==0? null: currentHolidayItem.getMonthUse());
+								&& currentHolidayItem.getMonthUse()==0? null: currentHolidayItem.getMonthUse());
 						if (currentMonth.compareTo(currentHolidayItem.getYm()) == 0) {
 							if (isRepresentSubstitute) {
 								// I4_3 代休_未消化_日数, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 								cells.get(rowIndexRepresentSubstitute, 10 + totalMonth)
 										.setValue(currentHolidayItem.getMonthExtinction() != null
-												&& currentHolidayItem.getMonthExtinction().intValue() == 0 ? null: currentHolidayItem.getMonthExtinction());
+												&& currentHolidayItem.getMonthExtinction() == 0 ? null: currentHolidayItem.getMonthExtinction());
 								if (currentHolidayItem.getMonthExtinction() != null
-										&& currentHolidayItem.getMonthExtinction().intValue() > 0) {
+										&& currentHolidayItem.getMonthExtinction() > 0) {
 									setForegroundRed(cells.get(rowIndexRepresentSubstitute, 10 + totalMonth));
 								}
 							}
@@ -736,7 +736,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 		if (isRepresentSubstitute) {
 			// I1_5 代休_未消化_日数
 			if (currentHolidayLeft.getMonthExtinction() != null
-					&& currentHolidayLeft.getMonthExtinction().intValue() > 0) {
+					&& currentHolidayLeft.getMonthExtinction() > 0) {
 				setForegroundRed(cells.get(firstRow, 8));
 			}
 		}
@@ -837,13 +837,13 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 					continue;
 				}
 				// J2_5 振休_発生, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
-				cells.get(firstRow, 10 + totalMonth).setValue(statusOfHDItem.getOccurredDay()!= null && statusOfHDItem.getOccurredDay().intValue()==0?null:statusOfHDItem.getOccurredDay() );
+				cells.get(firstRow, 10 + totalMonth).setValue(statusOfHDItem.getOccurredDay()!= null && statusOfHDItem.getOccurredDay()==0?null:statusOfHDItem.getOccurredDay() );
 				// J2_6 振休_使用, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
-				cells.get(firstRow + 1, 10 + totalMonth).setValue(statusOfHDItem.getUsedDays()!=null && statusOfHDItem.getUsedDays().intValue()==0?null:statusOfHDItem.getUsedDays());
+				cells.get(firstRow + 1, 10 + totalMonth).setValue(statusOfHDItem.getUsedDays()!=null && statusOfHDItem.getUsedDays()==0?null:statusOfHDItem.getUsedDays());
 				if (isUndigestedPause) {
 					// J2_7 振休_未消化, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
-					cells.get(rowIndexUndigestedPause, 10 + totalMonth).setValue(statusOfHDItem.getUnUsedDays()!=null &&statusOfHDItem.getUnUsedDays().intValue()==0?null: statusOfHDItem.getUnUsedDays());
-					if(statusOfHDItem.getUnUsedDays() !=null && statusOfHDItem.getUnUsedDays().intValue() > 0) {
+					cells.get(rowIndexUndigestedPause, 10 + totalMonth).setValue(statusOfHDItem.getUnUsedDays()!=null &&statusOfHDItem.getUnUsedDays()==0?null: statusOfHDItem.getUnUsedDays());
+					if(statusOfHDItem.getUnUsedDays() !=null && statusOfHDItem.getUnUsedDays() > 0) {
 						setForegroundRed(cells.get(rowIndexUndigestedPause, 10 + totalMonth));
 					}
 				}
@@ -865,17 +865,17 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 					if (maxRange >= totalMonth && totalMonth >= 0) {
 						// J2_5 振休_発生, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 						cells.get(firstRow, 10 + totalMonth).setValue(holidayRemainItem.getMonthOccurrence()!=null
-								&& holidayRemainItem.getMonthOccurrence().intValue()==0?null:holidayRemainItem.getMonthOccurrence());
+								&& holidayRemainItem.getMonthOccurrence()==0?null:holidayRemainItem.getMonthOccurrence());
 						// J2_6 振休_使用, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 						cells.get(firstRow + 1, 10 + totalMonth).setValue(holidayRemainItem.getMonthUse()!=null 
-								&& holidayRemainItem.getMonthUse().intValue()==0? null:holidayRemainItem.getMonthUse());
+								&& holidayRemainItem.getMonthUse()==0? null:holidayRemainItem.getMonthUse());
 						if (currentMonth.compareTo(holidayRemainItem.getYm()) == 0) {
 							if (isUndigestedPause) {
 								// J2_7 振休_未消化, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
 								cells.get(rowIndexUndigestedPause, 10 + totalMonth).setValue(holidayRemainItem.getMonthExtinction()!=null 
-										&& holidayRemainItem.getMonthExtinction().intValue()==0? null:holidayRemainItem.getMonthExtinction());
+										&& holidayRemainItem.getMonthExtinction()==0? null:holidayRemainItem.getMonthExtinction());
 								if(holidayRemainItem.getMonthExtinction() != null 
-										&& holidayRemainItem.getMonthExtinction().intValue() > 0) {
+										&& holidayRemainItem.getMonthExtinction() > 0) {
 									setForegroundRed(cells.get(rowIndexUndigestedPause, 10 + totalMonth));
 								}
 							}
@@ -916,7 +916,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 			if (isUndigestedPause) {
 				// J1_5 振休_未消化
 				if (currentHolidayRemainLeft.getMonthExtinction() != null
-						&& currentHolidayRemainLeft.getMonthExtinction().intValue() > 0) {
+						&& currentHolidayRemainLeft.getMonthExtinction() > 0) {
 					setForegroundRed(cells.get(firstRow, 8));
 				}
 			}
@@ -1020,7 +1020,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 							int totalMonth = totalMonths(dataSource.getStartMonth().yearMonth(), item.getYm());
 							if (maxRange >= totalMonth && totalMonth >= 0) {
 								// M2_5 特別休暇１_使用日数, set lại giá trị cho cột này nếu bằng 0 thì không hiển thị ra
-								cells.get(firstRow, 10 + totalMonth).setValue(item.getUseDays()!=null && item.getUseDays().intValue()==0? null:item.getUseDays());
+								cells.get(firstRow, 10 + totalMonth).setValue(item.getUseDays()!=null && item.getUseDays()==0? null:item.getUseDays());
 								// M2_7 特別休暇１_残数日数
 								cells.get(firstRow + 1, 10 + totalMonth).setValue(item.getRemainDays());
 								if (item.getRemainDays() < 0) {
@@ -1059,7 +1059,7 @@ public class HolidaysRemainingReportGeneratorImp extends AsposeCellsReportGenera
 						if (totalMonth <= maxRange && totalMonth >= 0) {
 							// M2_5 特別休暇１_使用日数
 							cells.get(firstRow, 10 + totalMonth).setValue(spVaCrurrentMonthImported.getUsedDate()!=null 
-									&& spVaCrurrentMonthImported.getUsedDate().intValue()==0?null:spVaCrurrentMonthImported.getUsedDate());
+									&& spVaCrurrentMonthImported.getUsedDate()==0?null:spVaCrurrentMonthImported.getUsedDate());
 							// M2_7 特別休暇１_残数日数
 //							cells.get(firstRow + 1, 10 + totalMonth).setValue(spVaCrurrentMonthImported.getRemainDate());
 							if (spVaCrurrentMonthImported.getRemainDate() < 0) {
